DMP-1717 Fixed some broken tests
